+++ src/trx/game/enum.c
@@ -179,6 +179,7 @@ static __attribute__((constructor)) void M_Init(void)
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_FLYBY_END, "flyby_end");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_CUTSCENE_TRIGGER, "cutscene_trigger");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_CUTSCENE_START, "cutscene_start");
+    ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_CUTSCENE_FRAME, "cutscene_frame");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_CUTSCENE_END, "cutscene_end");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_LEVEL_UNLOAD, "level_unload");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_UI_DRAW, "ui_draw");